arch: change variable types to match those of what is assigned to them
There are many places where "int len = strlen(foo);" is written, just for len to be passed as a parameter of size_t. This causes truncation and then expansion back from int to size_t. Not to mention this is poor logic and needless truncations can add extra unneeded instructions.

This patch aims to fix the worst offenders.

+++ arch/arm/vfp/vfpmodule.c
@@ -227,7 +227,7 @@ static void vfp_raise_sigfpe(unsigned int sicode, struct pt_regs *regs)
 
 static void vfp_panic(char *reason, u32 inst)
 {
-	int i;
+	unsigned i;
 
 	pr_err("VFP: Error: %s\n", reason);
 	pr_err("VFP: EXC 0x%08x SCR 0x%08x INST 0x%08x\n",
@@ -700,7 +700,7 @@ static struct undef_hook vfp_kmode_exception_hook[] = {{
 
 static int __init vfp_kmode_exception_hook_init(void)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < ARRAY_SIZE(vfp_kmode_exception_hook); i++)
 		register_undef_hook(&vfp_kmode_exception_hook[i]);
